refact(skymp5-functions-lib): refactor and add test cases (#728)
