+++ skymp5-functions-lib/index.ts
@@ -1,9 +1,10 @@
+import { SweetPieGameModeListener } from './src/logic/SweetPieGameModeListener';
+import { SweetPieMap } from './src/logic/SweetPieMap';
 import { MpApiInteractor } from './src/mpApiInteractor';
 import { BrowserProperty } from './src/props/browserProperty';
 import { ChatProperty } from './src/props/chatProperty';
 import { DialogProperty } from './src/props/dialogProperty';
 import { EvalProperty } from './src/props/evalProperty';
-import { SweetPieGameModeListener, SweetPieMap } from './src/sweetPie';
 import { LocationalData, Mp, PapyrusObject, PapyrusValue } from './src/types/mp';
 import { Timer } from './src/utils/timer';
