+++ skymp5-functions-lib/src/logic/GameModeListener.ts
@@ -1,0 +1,9 @@
+export interface GameModeListener {
+  onPlayerJoin?: (actorId: number) => void;
+  onPlayerLeave?: (actorId: number) => void;
+  onPlayerDeath?: (targetActorId: number, killerActorId?: number) => void;
+  everySecond?: () => void;
+  onPlayerChatInput?: (actorId: number, inputText: string, neighbors: number[], senderName: string) => void;
+  onPlayerDialogResponse?: (actorId: number, dialogId: number, buttonIndex: number) => void;
+  onPlayerActivateObject?: (casterActorId: number, targetObjectDesc: string, isTeleportDoor: boolean) => 'continue' | 'blockActivation';
+}
