+++ skymp5-functions-lib/src/logic/PlayerController.ts
@@ -1,0 +1,8 @@
+export type PlayerController = {
+  setSpawnPoint(player: number, pointName: string): void;
+  teleport(player: number, pointName: string): void;
+  showMessageBox(actorId: number, dialogId: number, caption: string, text: string, buttons: string[]): void;
+  sendChatMessage(actorId: number, text: string): void;
+  quitGame(actorId: number): void;
+  getName(actorId: number): string;
+}
