+++ skymp5-functions-lib/src/logic/SweetPieMap.ts
@@ -1,0 +1,7 @@
+export type SweetPieMap = {
+  safePointName: string;
+  mainSpawnPointName?: string;
+  safePlaceLeaveDoors?: string[];
+  safePlaceEnterDoors?: string[];
+  leaveRoundDoors?: string[];
+}
