+++ skymp5-functions-lib/src/mpApiInteractor.ts
@@ -1,7 +1,8 @@
+import { GameModeListener } from "./logic/GameModeListener";
+import { PlayerController } from "./logic/PlayerController";
 import { ChatProperty } from "./props/chatProperty";
 import { DialogProperty } from "./props/dialogProperty";
 import { EvalProperty } from "./props/evalProperty";
-import { GameModeListener, PlayerController } from "./sweetPie";
 import { Ctx } from "./types/ctx";
 import { LocationalData, Mp } from "./types/mp";
 import { PersistentStorage } from "./utils/persistentStorage";
